Moving XML Parser out of aws-c-auth and into aws-c-common
