+++ tests/CMakeLists.txt
@@ -202,15 +202,6 @@ add_test_case(sigv4_skip_xray_header_test)
 add_test_case(sigv4_skip_user_agent_header_test)
 add_test_case(sigv4_skip_custom_header_test)
 
-add_test_case(xml_parser_root_with_text)
-add_test_case(xml_parser_child_with_text)
-add_test_case(xml_parser_siblings_with_text)
-add_test_case(xml_parser_preamble_and_attributes)
-add_test_case(xml_parser_nested_node_same_name_test)
-add_test_case(xml_parser_nested_node_deep_recursion_test)
-add_test_case(xml_parser_too_many_attributes_test)
-add_test_case(xml_parser_name_too_long_test)
-
 add_test_case(sigv4_fail_date_header_test)
 add_test_case(sigv4_fail_content_header_test)
 add_test_case(sigv4_fail_authorization_header_test)
